file structure are changed and geo_proc class is added
I changed billiard detection file.
I implemented structure that uses class and
divided headers and detection codes.

I also added Geo_Proc class
Geo_Proc class helps to calculate 3D camera pose and relative pose of billiard.
