+++ app/src/main/cpp/basic.h
@@ -1,6 +1,7 @@
-#include <opencv2/opencv.hpp>
+#include <opencv2/core.hpp>
 #include <opencv2/highgui.hpp> 
 #include <opencv2/imgproc.hpp>
+#include <opencv2/calib3d.hpp>
 //#include <opencv2/features2d.hpp>
 //#include <opencv2/xfeatures2d.hpp>
 #include <iostream>
